LOL-AI/requirements-dev-linux.txt
alembic==1.8.1
anyio==3.6.1
attrs==22.1.0
aws-cfn-bootstrap==2.0
azure-ai-formrecognizer==3.2.0b2
azure-common==1.1.28
azure-core==1.25.1
backoff==1.11.1
boto3==1.24.79
botocore==1.27.79
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
databricks-cli==0.17.3
dill==0.3.5.1
docker==6.0.0
docopt==0.6.2
docutils==0.14
elastic-apm==6.12.0
elasticsearch==7.10.1
entrypoints==0.4
farm-haystack==1.8.0
fastapi==0.85.0
filelock==3.8.0
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
httptools==0.5.0
huggingface-hub==0.7.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflect==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
langdetect==1.0.9
lockfile==0.11.0
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
mlflow==1.29.0
mmh3==3.0.0
monotonic==1.6
more-itertools==8.14.0
msrest==0.7.1
networkx==2.6.3
nltk==3.7
num2words==0.5.12
numpy==1.21.6
oauthlib==3.2.1
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
posthog==2.1.2
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
protobuf==4.21.6
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
PyPDF2==2.10.9
pyrsistent==0.18.1
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
quantulum3==0.7.10
querystring-parser==1.2.4
rapidfuzz==2.10.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
simplejson==3.2.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
starlette==0.20.4
tabulate==0.8.10
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.12.1
torch==1.12.1+rocm5.1.1
torchvision==0.13.1+rocm5.1.1
tqdm==4.64.1
transformers==4.20.1
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.2.2
wordsegment==1.3.1
zipp==3.8.1
